refactor gulp utils
